https://github.com/toolkits for switch

+++ cpustat.go
@@ -1,6 +1,6 @@
 package sw
 
-func CpuUtilization(ip, community string, timeout int64) (int, error) {
+func CpuUtilization(ip, community string, timeout int) (int, error) {
 	vendor, err := SysVendor(ip, community, timeout)
 	method := "get"
 	var oid string
